*more work in progress *lots of internal gui control work *added isfs (nand) filesystem (read only) *compile fix for devkitPPC r24

+++ source/ImageOperations/ImageConverter.cpp
@@ -71,7 +71,7 @@ bool ImageConverter::LoadImage(const char * filepath)
 	SetOutPath(filepath);
 
 	u8 * buffer = NULL;
-	u64 filesize = 0;
+	u32 filesize = 0;
 
 	LoadFileToMemWithProgress(tr("Loading image..."), filepath, &buffer, &filesize);
